requirements.txt
fastapi
uvicorn
pillow
python-multipart
pydantic
numpy
pycairo
scikit-learn
# Optional/future modules:
torch
torchvision
tqdm
matplotlib
seaborn
jupyter
svgwrite 